Saber

+++ Saber/Saber.Node/Create.cs
@@ -219,7 +219,8 @@ public class Create : ClassCreate
 
         long count;
         count = this.NodeKind.Count;
-        int i;
+
+        long i;
         i = 0;
         while (i < count)
         {
